centralize fib-related API
